+++ src-tauri/tests/mod.rs
@@ -1,2 +1,2 @@
 pub mod common;
-pub mod tests;
+mod tests;
