+++ src-tauri/tests/tests/mod.rs
@@ -1,1 +1,2 @@
-pub mod load_dev_profile;
+mod load_dev_profile;
+mod qr_code_scanned;
